moving for common code point
