+++ cpp_common/host_data_receiver/pybind_install.sh
@@ -11,7 +11,6 @@ cmake ..
 make
 filename=$(find . -name "host_data_receiver*so")
 mv $filename host_data_receiver.so
-cp host_data_receiver.so \
-   ../../../../../../spinn_front_end_common/utility_models/
+cp host_data_receiver.so ../../spinn_front_end_common/common_host_binaries
 cd ../..
 rm -rf pybind11-demo
